pyproject.toml
[project]
name = "career-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
    "edge-tts",
    "speechrecognition",
    "pyttsx3",
    "playsound==1.2.2",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.0"

]
